Get or Set the property on the current holder (after load of extension)
An extension (and any web app for that matter) can replace objects in the DOM.
If this happens as part of lazy loading, it means that the property access
which triggered the lazy load, had its holder object replaced and the access
it thus being triggered on the original holder and not the current one (as
replaced by the extension).

This patch makes sure that we trigger it on the current holder.